fix id in user

+++ database/migrations/2020_10_25_120649_create_customer_infos_table.php
@@ -19,7 +19,7 @@ class CreateCustomerInfosTable extends Migration
             $table->string('email');
             $table->string('name');
             $table->string('phone_number');
-            $table->unsignedInteger('user_id');
+            $table->unsignedBigInteger('user_id');
             $table->foreign('user_id')->references('id')->on('users');
         });
     }
