+++ database/migrations/2020_10_27_090341_add_user_to_review.php
@@ -15,7 +15,7 @@ class AddUserToReview extends Migration
     {
         Schema::table('reviews', function (Blueprint $table) {
             //
-            $table->unsignedInteger('user_id')->nullable();
+            $table->unsignedBigInteger('user_id')->nullable();
             $table->foreign('user_id')->references('id')->on('users');
         });
     }
